move applyDragImage into ui/dnd
