+++ src/vs/base/browser/ui/list/list.css
@@ -60,30 +60,6 @@
 	outline: 0 !important;
 }
 
-/* Dnd */
-.monaco-drag-image {
-	display: inline-block;
-	padding: 1px 7px;
-	border-radius: 10px;
-	font-size: 12px;
-	position: absolute;
-	z-index: 1000;
-
-	background-color: var(--vscode-list-activeSelectionBackground);
-	color: var(--vscode-list-activeSelectionForeground);
-	outline: 1px solid var(--vscode-list-focusOutline);
-	outline-offset: -1px;
-
-	/*
-	 * Browsers apply an effect to the drag image when the div becomes too
-	 * large which makes them unreadable. Use max width so it does not happen
-	 */
-	max-width: 120px;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
 /* Filter */
 
 .monaco-list-type-filter-message {
